Split BB Track classes into separate files
