+++ include/BBTrack.h
@@ -27,75 +27,14 @@
 #ifndef BB_TRACK_H
 #define BB_TRACK_H
 
-#include <QtCore/QObject>
-#include <QtCore/QMap>
-#include <QStaticText>
 
-#include "TrackContentObjectView.h"
+#include <QtCore/QMap>
+
+#include "BBTCOView.h"
 #include "Track.h"
-#include "TrackView.h"
 
 class TrackLabelButton;
 class TrackContainer;
-
-
-class BBTCO : public TrackContentObject
-{
-public:
-	BBTCO( Track * _track );
-	virtual ~BBTCO() = default;
-
-	void saveSettings( QDomDocument & _doc, QDomElement & _parent ) override;
-	void loadSettings( const QDomElement & _this ) override;
-	inline QString nodeName() const override
-	{
-		return( "bbtco" );
-	}
-
-	int bbTrackIndex();
-
-	TrackContentObjectView * createView( TrackView * _tv ) override;
-
-private:
-
-
-	friend class BBTCOView;
-
-} ;
-
-
-
-class BBTCOView : public TrackContentObjectView
-{
-	Q_OBJECT
-public:
-	BBTCOView( TrackContentObject * _tco, TrackView * _tv );
-	virtual ~BBTCOView() = default;
-
-
-public slots:
-	void update() override;
-
-protected slots:
-	void openInBBEditor();
-	void resetName();
-	void changeName();
-
-
-protected:
-	void paintEvent( QPaintEvent * pe ) override;
-	void mouseDoubleClickEvent( QMouseEvent * _me ) override;
-	void constructContextMenu( QMenu * ) override;
-
-
-private:
-	BBTCO * m_bbTCO;
-	QPixmap m_paintPixmap;
-	
-	QStaticText m_staticTextName;
-} ;
-
-
 
 
 class LMMS_EXPORT BBTrack : public Track
@@ -149,34 +88,6 @@ private:
 	static infoMap s_infoMap;
 
 	friend class BBTrackView;
-
-} ;
-
-
-
-class BBTrackView : public TrackView
-{
-	Q_OBJECT
-public:
-	BBTrackView( BBTrack* bbt, TrackContainerView* tcv );
-	virtual ~BBTrackView();
-
-	bool close() override;
-
-	const BBTrack * getBBTrack() const
-	{
-		return( m_bbTrack );
-	}
-
-
-public slots:
-	void clickedTrackLabel();
-
-
-private:
-	BBTrack * m_bbTrack;
-	TrackLabelButton * m_trackLabel;
-
 } ;
